revision Resptioist

+++ routes/web.php
@@ -21,10 +21,11 @@ Route::resource('clients', 'ClientController')->except(['show'])->middleware('au
 Route::get('clients/getdata','ClientController@getdata')->name('clients.data');
 Route::get('/managers/getdata','ManagerController@getdata')->name('managers.data');
 Auth::routes();
-Route::get('/home', 'HomeController@index')->name('home')->middleware('auth','update');
+Route::get('/home', 'HomeController@index')->name('home')->middleware('auth','update','forbid-banned-user');
 Route::get('/error',function(){
     return view('errors.404');
 })->name('error');
+//
 
 //---receiptionist---//
 Route::get('receiptionists','ReceptionistController@index')->name('receiptionists.index')->middleware('auth','role:Admin|Manager','forbid-banned-user');
@@ -38,24 +39,24 @@ Route::delete('receiptionists/{id}', 'ReceptionistController@delete')->name('rec
 //--------------------------------------------------------------------------------------//
 
 //---floors--//
-Route::get('floors','FloorController@index')->name('floors.index');
-Route::get('floors/getdata','FloorController@getdata')->name('floors.data');
-Route::get('floors/create','FloorController@create')->name('floors.create');
-Route::post('floors/store','FloorController@store')->name('floors.store');
-Route::get('floors/{id}/edit', 'FloorController@edit')->name('floors.edit');
-Route::post('floors/{id}', 'FloorController@update')->name('floors.update');
-Route::delete('floors/{id}', 'FloorController@delete')->name('floors.delete');
+Route::get('floors','FloorController@index')->name('floors.index')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('floors/getdata','FloorController@getdata')->name('floors.data')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('floors/create','FloorController@create')->name('floors.create')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::post('floors/store','FloorController@store')->name('floors.store')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('floors/{id}/edit', 'FloorController@edit')->name('floors.edit')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::post('floors/{id}', 'FloorController@update')->name('floors.update')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::delete('floors/{id}', 'FloorController@delete')->name('floors.delete')->middleware('auth','role:Admin|Manager','forbid-banned-user');
 
 //--------------------------------------------------------//
 
 //rooms routes//
-Route::get('rooms','RoomController@index')->name('rooms.index');
-Route::get('rooms/getdata','RoomController@getdata')->name('rooms.data');
-Route::get('rooms/create','RoomController@create')->name('rooms.create');
-Route::post('rooms/store','RoomController@store')->name('rooms.store');
-Route::get('rooms/{id}/edit', 'RoomController@edit')->name('rooms.edit');
-Route::post('rooms/{id}', 'RoomController@update')->name('rooms.update');
-Route::delete('rooms/{id}', 'RoomController@delete')->name('rooms.delete');
+Route::get('rooms','RoomController@index')->name('rooms.index')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('rooms/getdata','RoomController@getdata')->name('rooms.data')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('rooms/create','RoomController@create')->name('rooms.create')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::post('rooms/store','RoomController@store')->name('rooms.store')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::get('rooms/{id}/edit', 'RoomController@edit')->name('rooms.edit')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::post('rooms/{id}', 'RoomController@update')->name('rooms.update')->middleware('auth','role:Admin|Manager','forbid-banned-user');
+Route::delete('rooms/{id}', 'RoomController@delete')->name('rooms.delete')->middleware('auth','role:Admin|Manager','forbid-banned-user');
 //--------------------------------------------------------//
